Updated product.h & productrelease.h
Moved sizes to public because other headers may need access to them for saving files. Updated both headers to have static functions that will handle the UI for their portion and will simply return their respective type. Also added an "isValid" method with which you can press "enter" in one of the UI functions and it will return an "invalid" type for the main to use to exit back to main menu.